pyproject.toml
[tool.poetry]
name = "Neuer Standard"
version = "0.1.0"
description = ""
authors = ["info@mskuibida.com", "DaniilMurai@gmail.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
alembic = "1.15.2"
annotated-types = "0.7.0"
anyio = "4.9.0"
asyncpg = "0.30.0"
bcrypt = "4.3.0"
cffi = "1.17.1"
click = "8.1.8"
colorama = "0.4.6"
cryptography = "44.0.2"
ecdsa = "0.19.1"
fastapi = "0.115.12"
greenlet = "3.2.1"
h11 = "0.14.0"
idna = "3.10"
Mako = "1.3.10"
markdown-it-py = "3.0.0"
MarkupSafe = "3.0.2"
mdurl = "0.1.2"
passlib = "1.7.4"
pyasn1 = "0.4.8"
pycparser = "2.22"
pydantic = "2.11.3"
pydantic-settings = "2.9.1"
pydantic_core = "2.33.1"
Pygments = "2.19.1"
python-dotenv = "1.1.0"
python-jose = "3.4.0"
rich = "14.0.0"
rsa = "4.9.1"
shellingham = "1.5.4"
six = "1.17.0"
sniffio = "1.3.1"
SQLAlchemy = "2.0.40"
starlette = "0.46.2"
typer = "0.15.2"
typing-inspection = "0.4.0"
typing_extensions = "4.13.2"
uvicorn = "0.34.2"
APScheduler = "^3.11.0"
pytz = "^2025.2"
pandas = "^2.3.0"
python-dateutil = "^2.9.0.post0"
redis = "^6.2.0"
python-i18n = "^0.3.9"
openpyxl = "^3.1.5"
httpx = "^0.28.1"
pytest = "^8.4.1"
aiofiles = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
